--- lib/screen/sign_up.dart
@@ -1,109 +1,0 @@
-import 'package:flutter/material.dart';
-
-import '../view/pages/loginhome/login_home_page.dart';
-
-class sign extends StatelessWidget {
-  const sign({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      decoration: BoxDecoration(
-        image: DecorationImage(
-          image: AssetImage('assets/reguster.png'),
-          fit: BoxFit.cover,
-        ),
-      ),
-      child: Scaffold(
-        backgroundColor: Colors.transparent,
-        body: SafeArea(
-          child: Column(
-            children: [
-              Padding(
-                padding: const EdgeInsets.symmetric(horizontal: 10),
-                child: Row(
-                  children: [
-                    IconButton(
-                      onPressed: () {
-                        Navigator.of(context)
-                            .push(MaterialPageRoute(builder: (_) => MyLogin()));
-                      },
-                      icon: Icon(Icons.arrow_back_ios),
-                      iconSize: 40,
-                    ),
-                    Text(
-                      '회원가입',
-                      style: TextStyle(
-                        fontSize: 30,
-                        fontWeight: FontWeight.bold,
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-              Container(
-                child: Padding(
-                  padding: EdgeInsets.only(top: 20, left: 40, right: 40),
-                  child: TextField(
-                    decoration: InputDecoration(
-                      fillColor: Colors.green,
-                      hintText: 'E-main',
-                      border: OutlineInputBorder(
-                        borderRadius: BorderRadius.circular(10),
-                        borderSide: BorderSide(width: 0.5),
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-              Container(
-                child: Padding(
-                  padding: EdgeInsets.only(top: 20, left: 40, right: 40),
-                  child: TextField(
-                    decoration: InputDecoration(
-                      fillColor: Colors.green,
-                      hintText: 'Password',
-                      border: OutlineInputBorder(
-                        borderRadius: BorderRadius.circular(10),
-                        borderSide: BorderSide(width: 0.5),
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-              Container(
-                child: Padding(
-                  padding: EdgeInsets.only(top: 20, left: 40, right: 40),
-                  child: TextField(
-                    decoration: InputDecoration(
-                      fillColor: Colors.green,
-                      hintText: 'Confirm Password',
-                      border: OutlineInputBorder(
-                        borderRadius: BorderRadius.circular(10),
-                        borderSide: BorderSide(width: 0.5),
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-              Expanded(
-                child: Container(
-                  alignment: Alignment.bottomCenter,
-                  padding: EdgeInsets.only(bottom: 30),
-                  child: ElevatedButton(
-                    onPressed: () {},
-                    child: Text('회원가입 완료'),
-                    style: ElevatedButton.styleFrom(
-                      backgroundColor: Colors.green,
-                      fixedSize: Size(300, 40),
-                    ),
-                  ),
-                ),
-              )
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
